Expand `Cloud.Announcement`
