+++ packages/core-typings/src/cloud/NpsSurveyAnnouncement.ts
@@ -1,0 +1,7 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+
+export interface NpsSurveyAnnouncement {
+	id: string;
+	startAt: Date;
+	expireAt: Date;
+}
